+++ js/app/routes.js
@@ -10,7 +10,8 @@ var
   About = require( "./components/About.react" ),
   UserPanel = require( "./components/UserPanel.react" ),
   RepositoryList = require( "./components/RepositoryList.react" ),
-  Profile = require( "./components/Profile.react" );
+  RepositoryDetail = require( "./components/RepositoryDetail.react" ),
+  Profile = require( "./components/Profile.react" ),
 
   routes = (
     <Router name="app" path="/" handler={KanbanerApp}>
@@ -20,6 +21,7 @@ var
       <Route name="user-panel" handler={UserPanel}>
         <DefaultRoute handler={RepositoryList} />
         <Route name="profile" handler={Profile} />
+        <Router name="repository-detail" path=":repositoryOwner/:repositoryName" handler={RepositoryDetail} />
       </Route>
     </Router>
   );