+++ js/app/stores/RepositoryStore.js
@@ -9,25 +9,44 @@ var
 
   REPOSITORY_LIST_CHANGED = "repository-list-changed",
   REPOSITORY_LIST_CHANGE_FAILED = "repository-list-update-failed",
+  REPOSITORY_SELECTED = "repository-selected",
+  REPOSITORY_SELECT_FAILED = "repository-select-failed",
 
   repositoryList = [],
+  selectedRepository = null,
 
   repositoryListChanged  = function( data ) {
     repositoryList = data;
-    RepositoryListStore.emitRepositoryListChanged();
+    RepositoryStore.emitRepositoryListChanged();
   },
 
   repositoryListChangeFailed = function() {
-    RepositoryListStore.emitRepositoryListChangeFailed();
+    RepositoryStore.emitRepositoryListChangeFailed();
   },
 
-  RepositoryListStore = assign({}, EventEmitter.prototype, {
+  repositorySelected = function ( data ) {
+    selectedRepository = data;
+    RepositoryStore.emitRepositorySelectedEvent();
+  },
+
+  repositorySelectFailed = function () {
+    RepositoryStore.emitRepositorySelectFailedEvent();
+  },
+
+  RepositoryStore = assign({}, EventEmitter.prototype, {
     /**
      * Return list of available repositories
      * @returns {Array}
      */
     getRepositoryList: function() {
       return repositoryList;
+    },
+
+    /**
+     * Return selected repository
+     */
+    getSelectedRepository: function() {
+      return selectedRepository;
     },
 
     /**
@@ -74,11 +93,57 @@ var
      */
     emitRepositoryListChangeFailed: function() {
       this.emit( REPOSITORY_LIST_CHANGE_FAILED );
+    },
+
+
+    /**
+     * Add callback to listen REPOSITORY_SELECT event
+     * @param {function} callback
+     */
+    addRepositorySelectedListener: function( callback ) {
+      this.on( REPOSITORY_SELECTED, callback );
+    },
+
+    /**
+     * Remove callback from listen REPOSITORY_SELECT event
+     */
+    removeRepositorySelectedListener: function( callback ) {
+      this.removeListener( REPOSITORY_SELECTED, callback );
+    },
+
+    /**
+     * Send REPOSITORY_SELECTED event
+     */
+    emitRepositorySelectedEvent: function() {
+      this.emit( REPOSITORY_SELECTED )
+    },
+
+    /**
+     * Add callback to listen REPOSITORY_SELECT_FAILED event
+     * @param {function} callback
+     */
+    addRepositorySelectFailedListener: function( callback ) {
+      this.on( REPOSITORY_SELECT_FAILED, callback );
+    },
+
+    /**
+     * Remove callback from listen REPOSITORY_SELECTED_FAILED event
+     * @param {function} callback
+     */
+    removeRepositorySelectFailedListener: function( callback ) {
+      this.removeListener( REPOSITORY_SELECT_FAILED, callback);
+    },
+
+    /**
+     * Send REPOSITORY_SELECT_FAILED event
+     */
+    emitRepositorySelectFailedEvent: function() {
+      this.emit( REPOSITORY_SELECT_FAILED );
     }
 
   });
 
-  RepositoryListStore.dispatchToken = KanbanerDispatcher.register(
+  RepositoryStore.dispatchToken = KanbanerDispatcher.register(
     function(payload) {
       var action = payload.action;
 
@@ -89,8 +154,14 @@ var
         case ActionTypes.REPOSITORIES_LOAD_FAIL:
           repositoryListChangeFailed();
           break;
+        case ActionTypes.REPOSITORY_SELECT_SUCCESS:
+          repositorySelected( action.data );
+          break;
+        case ActionTypes.REPOSITORY_SELECT_FAIL:
+          repositorySelectFailed();
+          break;
       }
     }
   );
 
-module.exports = RepositoryListStore;
+module.exports = RepositoryStore;